# perf: Improve performance of computing number of not-cloned repos - Part 2