--- a/cmd/gitserver/server/list.go
+++ b/cmd/gitserver/server/list.go
@@ -5,6 +5,8 @@ import (
 	"net/http"
 	"os"
 	"path/filepath"
+
+	"github.com/karrick/godirwalk"
 )
 
 func (s *Server) handleList(w http.ResponseWriter, r *http.Request) {
@@ -23,54 +25,58 @@ func (s *Server) handleList(w http.ResponseWriter, r *http.Request) {
 		return
 
 	case query("cloned"):
-		err := filepath.Walk(s.ReposDir, func(path string, info os.FileInfo, err error) error {
-			if ctx.Err() != nil {
-				return ctx.Err()
-			}
+		err := godirwalk.Walk(s.ReposDir, &godirwalk.Options{
+			Callback: func(path string, de *godirwalk.Dirent) error {
+				if ctx.Err() != nil {
+					return ctx.Err()
+				}
 
-			if s.ignorePath(path) {
-				if info.IsDir() {
+				if s.ignorePath(path) {
+					if de.IsDir() {
+						return filepath.SkipDir
+					}
+					return nil
+				}
+
+				// We only care about directories
+				if !de.IsDir() {
+					return nil
+				}
+
+				// New style git directory layout
+				if filepath.Base(path) == ".git" {
+					name, err := filepath.Rel(s.ReposDir, filepath.Dir(path))
+					if err != nil {
+						return err
+					}
+					repos = append(repos, name)
 					return filepath.SkipDir
 				}
-				return nil
-			}
 
-			if err != nil {
-				return nil
-			}
+				// For old-style directory layouts we need to do an extra extra
+				// stat to check if this is a repo.
+				if _, err := os.Stat(filepath.Join(path, "HEAD")); os.IsNotExist(err) {
+					// HEAD doesn't exist, so keep recursing
+					return nil
+				} else if err != nil {
+					return err
+				}
 
-			// We only care about directories
-			if !info.IsDir() {
-				return nil
-			}
-
-			// New style git directory layout
-			if filepath.Base(path) == ".git" {
-				name, err := filepath.Rel(s.ReposDir, filepath.Dir(path))
+				// path is an old style git repo since it contains HEAD
+				name, err := filepath.Rel(s.ReposDir, path)
 				if err != nil {
 					return err
 				}
 				repos = append(repos, name)
 				return filepath.SkipDir
-			}
-
-			// For old-style directory layouts we need to do an extra extra
-			// stat to check if this is a repo.
-			if _, err := os.Stat(filepath.Join(path, "HEAD")); os.IsNotExist(err) {
-				// HEAD doesn't exist, so keep recursing
-				return nil
-			} else if err != nil {
-				return err
-			}
-
-			// path is an old style git repo since it contains HEAD
-			name, err := filepath.Rel(s.ReposDir, path)
-			if err != nil {
-				return err
-			}
-			repos = append(repos, name)
-			return filepath.SkipDir
+			},
+			ErrorCallback: func(path string, err error) godirwalk.ErrorAction {
+				// Ignore errors and simply continue with other nodes
+				return godirwalk.SkipNode
+			},
+			Unsorted: true,
 		})
+
 		if err != nil {
 			http.Error(w, err.Error(), http.StatusInternalServerError)
 			return